utils/ml_models.py
import pandas as pd
import numpy as np
from sklearn.ensemble import RandomForestRegressor
import streamlit as st

class SavingsRecommender:
    def __init__(self):
        self.model = None
        
    def get_recommendation(self, user_data):
        # Logique simplifiée de recommandation
        savings_advice = [
            "Ouvrez un compte épargne à 3% pour maximiser vos rendements",
            "Épargnez 20% de votre revenu mensuel",
            "Pensez à l'épargne retraite avec notre produit dédié",
            "Diversifiez avec des fonds d'investissement"
        ]
        
        return np.random.choice(savings_advice)

class CreditPredictor:
    def __init__(self):
        pass
        
    def predict_eligibility(self, user_data):
        # Simulation d'éligibilité au crédit
        return {
            'eligible': True,
            'montant_max': 50000,
            'taux_interet': 5.2,
            'duree_max': 60
        }