requirements.txt
streamlit==1.28.0
pytesseract==0.3.10
opencv-python==4.8.1.78
pillow==10.0.1
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
plotly==5.15.0
python-dotenv==1.0.0
requests==2.31.0
transformers==4.34.0
torch==2.0.1
streamlit-chat==0.1.1
python-multipart==0.0.6